feat: added world.Lock/Unlock + args check for debug
Added WorldAssertActions to config.go. This variable controls the
verification of extra states when adding/removing entities/components
and output the problem if LogEnabled = true (for example, trying to use
an entity or accessing its components after it was removed)

Added Lock/Unlock functions to World. It allows the insertion or removal
of entities and components inside Filter loops. The actions are
registered and the changes are added to world when Unlock is called.
Note that you need to call Unlock the same ammount of times you call
Lock, otherwise the world will continue to be in locked state.